get old style PSF photometry working
